script rename for dataloaders
